+++ Assembly.Projecto.Final.Services/Services/AddressService.cs
@@ -23,30 +23,64 @@ namespace Assembly.Projecto.Final.Services.Services
 
         public AddressDto Add(CreateAddressDto createAddressDto)
         {
-            return _unitOfWork.AddressRepository.Add(address);
+            var address = _mapper.Map<Address>(createAddressDto);
+
+            Address addedAddress;
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                addedAddress = _unitOfWork.AddressRepository.Add(address);
+
+                _unitOfWork.Commit(); 
+            }
+
+            return _mapper.Map<AddressDto>(addedAddress);
         }
 
-        public Address Delete(Address address)
+        public AddressDto Delete(AddressDto addressDto)
         {
-            return _unitOfWork.AddressRepository.Delete(address);
+            var address = _mapper.Map<Address>(addressDto);
+
+            Address deletedAddress;
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAddress = _unitOfWork.AddressRepository.GetById(address.Id);
+
+                if (foundedAddress == null) 
+                {
+                    throw new ArgumentNullException(nameof(foundedAddress), "Not found");
+                }
+
+                deletedAddress =_unitOfWork.AddressRepository.Delete(address);
+
+                _unitOfWork.Commit();
+            }
+
+
+            return _mapper.Map<AddressDto>(deletedAddress);
         }
 
-        public Address? Delete(int id)
+        public AddressDto Delete(int id)
         {
             return _unitOfWork.AddressRepository.Delete(id);
         }
 
-        public List<Address> GetAll()
+        public List<AddressDto> GetAll()
         {
             return _unitOfWork.AddressRepository.GetAll();
         }
 
-        public Address? GetById(int id)
+        public AddressDto GetById(int id)
         {
             return _unitOfWork.AddressRepository.GetById(id);
         }
 
-        public Address Update(Address address)
+        public AddressDto Update(AddressDto addressDto)
         {
              return _unitOfWork.AddressRepository.Update(address);
         }
